perf: optimize DSP pipeline for ESP32
- FFT: pre-compute Hamming window at construction, eliminating ~88K
  cosine evaluations/sec from the hot path
- Math: replace std::sqrt/std::exp with sqrtf/expf throughout DSP
  headers to guarantee single-precision on Xtensa (no hardware double FPU)
- OnsetDetector: replace std::vector with fixed circular arrays and
  incremental sum/sq_sum for O(1) threshold computation; zero heap
  allocations in the processing path
- Shared data: replace portMUX spinlock with lock-free double-buffer
  and volatile index swap between FFT task and main loop
- FFT task: replace 1ms polling loop with xTaskNotifyGive/Take,
  eliminating ~900 wasted task switches/sec
- Ring buffer: shrink from 2048 to 1024 elements, saving 4 KB SRAM

+++ components/audio_reactive/beat_tracker.h
@@ -34,7 +34,7 @@ class BeatTracker {
         float rp = update_rate_hz * 60.0f / 120.0f;
         for (size_t i = 0; i < MAX_LAG; i++) {
             float x = static_cast<float>(i + 1);
-            rayleigh_[i] = (x / (rp * rp)) * std::exp(-0.5f * x * x / (rp * rp));
+            rayleigh_[i] = (x / (rp * rp)) * expf(-0.5f * x * x / (rp * rp));
         }
 
         std::memset(df_buffer_, 0, sizeof(df_buffer_));
